@hahnjo 2024-01-31
> > I believe you need the "clean build" label here to avoid incremental builds tripping over the old pcms...
> 
> I am lost -- was only triggering a rebuild enough? If so, then I find it hard to believe we cannot compile root without these patches.

No you need a new workflow run (just push an empty commit or amend the last to get a new commit hash) - the state of labels is recorded at the beginning to avoid some builds running with different settings than others