@hahnjo 2024-02-02
> @hahnjo, according to our ci this patch seems fine.

IIUC `LoadExternalSpecializationsLazily` is still off by default, right? But good that it passes now in this off-by-default mode :+1: