@smuzaffar 2024-02-02
> @smuzaffar, can we run this PR against cmssw?

CMSSW tests are running via https://github.com/cms-sw/root/pull/198 (note that cmssw tests are running for non-CXXMODULES builds)