@vgvassilev 2024-02-03
> > @smuzaffar, can we run this PR against cmssw?
> 
> CMSSW tests are running via [cms-sw#198](https://github.com/cms-sw/root/pull/198) (note that cmssw tests are running for non-CXXMODULES builds)

@smuzaffar, can you interpret the results?