@ChuanqiXu9 2024-03-01
It looks like we need to remove the preifx `llvm::endianness` from `llvm::endianness::little` in llvm16.