@ChuanqiXu9 2024-03-01
Wow, this looks worse than my original patch. But it shows these errors are build time errors instead of runtime errors. It is slightly odd to me. 

BTW, when I try to run root locally, I meet some odd problems and it looks like I can proceed after cleanning the cache.