@vgvassilev 2024-03-05
> > > Wow, this looks worse than my original patch. But it shows these errors are build time errors instead of runtime errors. It is slightly odd to me.
> > 
> > 
> > Yes, it fails during ROOT's build but already in a stage where we are running some built tools to build our modules. This involves loading modules of dependent libraries, so it will trigger the template specialization loading logic.
> 
> Got it. Thanks.
> 
> @vgvassilev It turns out that the reason for so many failures is we have an ExternalASTSource in root out of the llvm tree, so we don't update that. [ChuanqiXu9@570fd78](https://github.com/ChuanqiXu9/root/commit/570fd783875671d346c7cdc0d98a8a9afcad05a4)
> 
> There is also another bug about the order to remove iterators: [ChuanqiXu9@7fb4aa9](https://github.com/ChuanqiXu9/root/commit/7fb4aa9d11c6966b54745009187928dd13cec566)
> 
> Now I can build the project in https://github.com/ChuanqiXu9/root/commits/chuanqi_pr83233/. I only run `make` instead of `ctest`. But it may be fine as we said the logic happens at build time actually. I think you can push the 2 commits to the branch to test it with the CI again.

@ChuanqiXu9, that is really awesome! Apologies for the oversight of the ExternalASTSource. I've pushed these changes and let's see what happens.