@vgvassilev 2024-03-05
> @vgvassilev FYI `mac14arm` fails to build the dictionary for `TMVAUtils`, with a crash in `cling::DeclCollector::HandleInterestingDecl`...

Ha, ok, until now I only saw `Failed to open the S3 connection: You must provide an auth secret.`. That might be "just" missing a `PushTransactionRAII` on our side. Is there a way to build a debug + asserts version of that?