@ChuanqiXu9 2024-03-05
> > > > @vgvassilev Great!
> > > > For performances, it looks like you lack a commit from the branch https://github.com/ChuanqiXu9/root/commits/chuanqi_pr83233/, which tries to reduce the time we calculating the ODR Hash.
> > > 
> > > 
> > > Do you mean this one: [ChuanqiXu9@2467fe7](https://github.com/ChuanqiXu9/root/commit/2467fe73721150e3b70bc17cb36e03507257c6e5) ?
> > 
> > 
> > Yes
> 
> @ChuanqiXu9, looks like the commit I took is broken. Did you forget to push your latest changes?

Weird. It compiles good locally. I don't know why it can work in my local environment as it accesses the protected type actually. I just tried to refactor this as the newest commit in https://github.com/ChuanqiXu9/root/commits/chuanqi_pr83233/