@vgvassilev 2024-03-06
@ChuanqiXu9, this looks good on our side. Performance is also within the same ballpark. If I run `/usr/bin/time -v root.exe -l -b -q tutorials/hsimple.C ` I get:

| | master    | this pr | this PR w/o  [ChuanqiXu9@2467fe7](https://github.com/ChuanqiXu9/root/commit/2467fe73721150e3b70bc17cb36e03507257c6e5)
|--------|----------|:-------------:|------:|
user time: |  0.53s | 0.50s| 0.49s |
|rss: | 309436 | 310196 | 310192|

@smuzaffar, can you test this PR against cmssw?